Culling now works properly, fixed unit tests
Fixed issue with culling not working, which had to do with the fact that the indice data was wrong for each direction. Each Chunk now tracks how many vertices and indices it is currently using. Changed direction naming system to cardinal units. Fixed unit tests assert statements, as actual and expectation were swapped.